fix: client
M  public/locales/en/common.json
M  src/app/(common)/_profile/_hooks/ProfileHook.jsx
D  src/app/api/doctoral-center/manager-expert/unauthorized/group/route.js
M  src/app/api/doctoral-center/manager-expert/unauthorized/route.js
M  src/app/doctoral-center/(manager-expert)/_hooks/UnauthorizedUsersHook.jsx
M  src/lib/api/DoctoralCenterAPI.jsx
