+++ public/locales/en/common.json
@@ -103,6 +103,7 @@
   "Експерт": "expert",
   "Ръководител": "manager",
   "Докторант": "phd",
+  "докторант": "phd",
   "Докторантски център": "Doctoral center",
   "Член от комитета": "committee",
   "Преподавател": "teacher",
