--- src/app/api/doctoral-center/manager-expert/unauthorized/group/route.js
@@ -1,13 +1,0 @@
-import ServerRoute from "@/lib/api/router/server";
-import { url } from "../../url";
-
-const { route } = ServerRoute();
-
-export async function POST(request) {
-  return await route({
-    url: url + "/unauthorized-users/group",
-    method: "POST",
-    request: request,
-    queryParams: ["group"]
-  });
-}
